Refactor place and route
* Move place-and-route tests inline with code to match rest of project.
* Fix PEP-8 complaints about place-and-route.
* Break out generic geometry functions into rig.geometry module.
* Improve documentation and move tutorial-like sections into sphinx.
