+++ docs/source/conf.py
@@ -30,6 +30,7 @@ import os   # noqa
 # ones.
 extensions = [
     'sphinx.ext.autodoc',
+    'sphinx.ext.intersphinx',
     'numpydoc',
 ]
 
@@ -48,6 +49,9 @@ master_doc = 'index'
 # General information about the project.
 project = u'Rig'
 copyright = u'2015, Andrew Mundy, Jonathan Heathcote'
+
+# Autodoc should list members in the same order as the source code.
+autodoc_member_order = "bysource"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -95,6 +99,9 @@ pygments_style = 'sphinx'
 
 # If true, keep warnings as "system message" paragraphs in the built documents.
 #keep_warnings = False
+
+# Search Python docs for extra definitions.
+intersphinx_mapping = {'python': ('http://docs.python.org/3.2', None)}
 
 
 # -- Options for HTML output ----------------------------------------------
