+++ docs/source/geometry.rst
@@ -1,0 +1,5 @@
+Machine geometry utility functions
+==================================
+
+.. automodule:: rig.geometry
+    :members:
