+++ docs/source/machine.rst
@@ -1,0 +1,5 @@
+Machine resource models
+-----------------------
+
+.. automodule:: rig.machine
+    :members:
