+++ docs/source/place_and_route/allocation_algorithms.rst
@@ -1,0 +1,5 @@
+Allocation algorithms
+---------------------
+
+.. autofunction:: rig.place_and_route.allocate.greedy.allocate
+
