+++ docs/source/place_and_route/placement_algorithms.rst
@@ -1,0 +1,4 @@
+Placement algorithms
+--------------------
+
+.. autofunction:: rig.place_and_route.place.hilbert.place
