+++ docs/source/place_and_route/routing_algorithms.rst
@@ -1,0 +1,6 @@
+Routing algorithms
+------------------
+
+.. autofunction:: rig.place_and_route.route.ner.route
+
+
