+++ docs/source/routing_table.rst
@@ -1,0 +1,6 @@
+Routing table data structures
+-----------------------------
+
+.. automodule:: rig.routing_table
+    :members:
+
